+++ src/Bridges/NetteRobotLoader/TestSuitesFinder.php
@@ -4,6 +4,7 @@ declare(strict_types=1);
 namespace MyTester\Bridges\NetteRobotLoader;
 
 use MyTester\BaseTestSuitesFinder;
+use MyTester\TestSuitesSelectionCriteria;
 use Nette\Loaders\RobotLoader;
 use Nette\Utils\FileSystem;
 
@@ -15,18 +16,19 @@ use Nette\Utils\FileSystem;
  */
 final class TestSuitesFinder extends BaseTestSuitesFinder
 {
-    public function getSuites(string $folder): array
+    public function getSuites(TestSuitesSelectionCriteria $criteria): array
     {
         if (!$this->isAvailable()) {
             return [];
         }
+        $folder = $criteria->testsFolderProvider->folder;
         $suites = [];
         $robot = new RobotLoader();
         $tempDir = "$folder/temp/cache/Robot.Loader";
         FileSystem::createDir($tempDir);
         $robot->setTempDirectory($tempDir);
         $robot->addDirectory($folder);
-        $robot->acceptFiles = ["*" . self::FILENAME_SUFFIX, ];
+        $robot->acceptFiles = ["*" . $criteria->filenameSuffix, ];
         $robot->rebuild();
         $robot->register();
         $classes = $robot->getIndexedClasses();