+++ src/ComposerTestSuitesFinder.php
@@ -12,15 +12,15 @@ namespace MyTester;
  */
 final class ComposerTestSuitesFinder extends BaseTestSuitesFinder
 {
-    public function getSuites(string $folder): array
+    public function getSuites(TestSuitesSelectionCriteria $criteria): array
     {
         $suites = [];
-        $folder = (string) realpath($folder);
+        $folder = (string) realpath($criteria->testsFolderProvider->folder);
         /** @var array<class-string, string> $classMap */
         $classMap = require \findVendorDirectory() . "/composer/autoload_classmap.php";
         foreach ($classMap as $class => $file) {
             $file = (string) realpath($file);
-            if (!str_starts_with($file, $folder) || !str_ends_with($file, self::FILENAME_SUFFIX)) {
+            if (!str_starts_with($file, $folder) || !str_ends_with($file, $criteria->filenameSuffix)) {
                 continue;
             }
             if ($this->isTestSuite($class)) {