+++ tests/Bridges/NetteRobotLoader/TestSuitesFinderTest.php
@@ -5,6 +5,8 @@ namespace MyTester\Bridges\NetteRobotLoader;
 
 use MyTester\Attributes\TestSuite;
 use MyTester\TestCase;
+use MyTester\TestsFolderProvider;
+use MyTester\TestSuitesSelectionCriteria;
 
 /**
  * Test suite for class TestSuitesFinder
@@ -17,7 +19,9 @@ final class TestSuitesFinderTest extends TestCase
     public function testGetSuites(): void
     {
         $testSuitesFinder = new TestSuitesFinder();
-        $suites = $testSuitesFinder->getSuites(__DIR__ . "/../../");
+        $suites = $testSuitesFinder->getSuites(
+            new TestSuitesSelectionCriteria(new TestsFolderProvider(__DIR__ . "/../../"))
+        );
         $this->assertCount(38, $suites);
     }
 }